requirements.txt
PyYAML
rich
python-i18n
python3-nmap
pyside6-addons
requests